Fix/cleanup

+++ recipes-core/images/openvario-recovery-image_2.0.bb
@@ -6,6 +6,8 @@ LIC_FILES_CHKSUM = "file://${COMMON_LICENSE_DIR}/MIT;md5=0835ade698e0bcf8506ecda
 
 require ov-revision.inc
 
+inherit deploy
+
 S = "${WORKDIR}/${PN}-${PV}"
 
 SRC_URI = "\
@@ -20,6 +22,8 @@ DEPENDS = "\
         u-boot \
     "
 
+do_configure[depends] += "virtual/kernel:do_deploy openvario-recovery-initramfs:do_image_complete"
+
 # DEPENDS += "ov-revision.inc "
 
 do_compile[deptask] = "do_rm_work"
@@ -29,8 +33,14 @@ do_configure () {
 	cp ${WORKDIR}/openvario-recovery.its ${S}
 	#new image
 	dd if=${DEPLOY_DIR_IMAGE}/uImage of=${S}/Image bs=64 skip=1
-	# new initramfs (TBR: added rootfs into name)
-	cp -v ${DEPLOY_DIR_IMAGE}/openvario-base-initramfs-${MACHINE}.rootfs.cpio.gz ${S}/initramfs.cpio.gz
+    # Accept both the stable deploy symlink and the versioned rootfs artifact name.
+    if [ -f ${DEPLOY_DIR_IMAGE}/openvario-base-initramfs-${MACHINE}.rootfs.cpio.gz ]; then
+        cp -v ${DEPLOY_DIR_IMAGE}/openvario-base-initramfs-${MACHINE}.rootfs.cpio.gz ${S}/initramfs.cpio.gz
+    elif [ -f ${DEPLOY_DIR_IMAGE}/openvario-base-initramfs-${MACHINE}.cpio.gz ]; then
+        cp -v ${DEPLOY_DIR_IMAGE}/openvario-base-initramfs-${MACHINE}.cpio.gz ${S}/initramfs.cpio.gz
+    else
+        bbfatal "Missing initramfs artifact for ${MACHINE} in ${DEPLOY_DIR_IMAGE}"
+    fi
         cp -v ${DEPLOY_DIR_IMAGE}/${MACHINE}.dtb ${S}/openvario.dtb
 }
 
@@ -48,6 +58,8 @@ do_mkimage () {
 
 addtask mkimage after do_configure before do_install
 
-do_install () {
-    cp ${S}/ov-recovery.itb ${DEPLOY_DIR_IMAGE}
+do_deploy() {
+    install -Dm0644 ${S}/ov-recovery.itb ${DEPLOYDIR}/ov-recovery.itb
 }
+
+addtask deploy after do_mkimage before do_build
